Move tests to o.e.swtchart.test package
Reduces "split package" issues between test fragment and bundle.
