+++ org.eclipse.swtchart.test/src/org/eclipse/swtchart/test/ErrorBarTest.java
@@ -10,7 +10,7 @@
  * Contributors:
  * yoshitaka - initial API and implementation
  *******************************************************************************/
-package org.eclipse.swtchart;
+package org.eclipse.swtchart.test;
 
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -21,9 +21,12 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.widgets.Display;
+import org.eclipse.swtchart.IErrorBar;
+import org.eclipse.swtchart.ISeries;
+import org.eclipse.swtchart.Range;
 import org.eclipse.swtchart.IErrorBar.ErrorBarType;
 import org.eclipse.swtchart.ISeries.SeriesType;
-import org.eclipse.swtchart.util.ChartTestCase;
+import org.eclipse.swtchart.test.util.ChartTestCase;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
