+++ org.eclipse.swtchart.test/src/org/eclipse/swtchart/test/testmodel/BusinessObject.java
@@ -8,7 +8,7 @@
  * Contributors:
  * Christoph Läubrich - initial API and implementation
  *******************************************************************************/
-package org.eclipse.swtchart.testmodel;
+package org.eclipse.swtchart.test.testmodel;
 
 import java.util.Date;
